relocate DefaultTimerDelegate to platform
