+++ examples/lighting-app/nrfconnect/prj.conf
@@ -57,3 +57,5 @@ CONFIG_USE_SEGGER_RTT=n
 # Enable Factory Data feature
 CONFIG_CHIP_FACTORY_DATA=y
 CONFIG_CHIP_FACTORY_DATA_BUILD=y
+
+# CONFIG_OPENTHREAD_SOURCES=y
