+++ src/app/tests/TestCommissioningWindowManager.cpp
@@ -15,7 +15,6 @@
  *    limitations under the License.
  */
 
-#include <app/DefaultTimerDelegate.h>
 #include <app/TestEventTriggerDelegate.h>
 #include <app/reporting/ReportSchedulerImpl.h>
 #include <app/server/CommissioningWindowManager.h>
@@ -28,6 +27,7 @@
 #include <platform/CHIPDeviceLayer.h>
 #include <platform/CommissionableDataProvider.h>
 #include <platform/ConfigurationManager.h>
+#include <platform/DefaultTimerDelegate.h>
 #include <platform/PlatformManager.h>
 #include <platform/TestOnlyCommissionableDataProvider.h>
 #include <protocols/secure_channel/PASESession.h>
