+++ src/controller/CHIPDeviceControllerFactory.cpp
@@ -24,13 +24,13 @@
 
 #include <controller/CHIPDeviceControllerFactory.h>
 
-#include <app/DefaultTimerDelegate.h>
 #include <app/InteractionModelEngine.h>
 #include <app/OperationalSessionSetup.h>
 #include <app/reporting/ReportSchedulerImpl.h>
 #include <app/util/DataModelHandler.h>
 #include <lib/core/ErrorStr.h>
 #include <messaging/ReliableMessageProtocolConfig.h>
+#include <platform/DefaultTimerDelegate.h>
 
 #if CONFIG_DEVICE_LAYER
 #include <platform/CHIPDeviceLayer.h>
